# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Polo Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clubs/{clubId}/fields:
    get:
      operationId: listFields
      tags: [fields]
      summary: List club fields
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fields list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Field"
    post:
      operationId: createField
      tags: [fields]
      summary: Create a field
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFieldRequest"
      responses:
        "201":
          description: Field created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
  /fields/{fieldId}:
    put:
      operationId: updateField
      tags: [fields]
      summary: Update a field
      parameters:
        - name: fieldId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFieldRequest"
      responses:
        "200":
          description: Field updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
    delete:
      operationId: deleteField
      tags: [fields]
      summary: Delete a field
      parameters:
        - name: fieldId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Field deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /invites/accept:
    post:
      operationId: acceptInvite
      tags: [auth]
      summary: Accept an invite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInviteRequest"
      responses:
        "200":
          description: Invite accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /invites/{token}:
    get:
      operationId: getInviteByToken
      tags: [auth]
      summary: Get invite details by token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteDetail"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserWithRoles"
        token:
          type: string
      required:
        - user
        - token
    UserWithRoles:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [spectator, team_manager, admin]
        clubMemberships:
          type: array
          items:
            $ref: "#/components/schemas/ClubMembershipBrief"
        teamAssignments:
          type: array
          items:
            $ref: "#/components/schemas/TeamAssignmentBrief"
      required:
        - id
        - email
        - displayName
        - role
        - clubMemberships
        - teamAssignments
    ClubMembershipBrief:
      type: object
      properties:
        clubId:
          type: string
          format: uuid
        clubName:
          type: string
        clubSlug:
          type: string
        role:
          type: string
          enum: [owner, manager]
      required:
        - clubId
        - clubName
        - clubSlug
        - role
    TeamAssignmentBrief:
      type: object
      properties:
        assignmentId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        teamName:
          type: string
        tournamentId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
      required:
        - assignmentId
        - teamId
        - teamName
        - status
...
    Field:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        number:
          type: integer
          nullable: true
        lat:
          type: string
          nullable: true
        lng:
          type: string
          nullable: true
        surfaceType:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - id
        - clubId
    CreateFieldRequest:
      type: object
      properties:
        name:
          type: string
        number:
          type: integer
        lat:
          type: string
        lng:
          type: string
        surfaceType:
          type: string
        isActive:
          type: boolean
    UpdateFieldRequest:
      type: object
      properties:
        name:
          type: string
        number:
          type: integer
        lat:
          type: string
        lng:
          type: string
        surfaceType:
          type: string
        isActive:
          type: boolean
...
    AcceptInviteRequest:
      type: object
      properties:
        token:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        displayName:
          type: string
      required:
        - token
    InviteDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        type:
          type: string
          enum: [team_manager, club_manager]
        teamName:
          type: string
          nullable: true
        clubName:
          type: string
          nullable: true
        expired:
          type: boolean
        used:
          type: boolean
      required:
        - id
        - email
        - type
        - expired
        - used
...